test(draft): extract shared test fixtures into draft.fixtures.ts
draft.service_test.ts had grown to ~2500 lines, much of it shared mock
data and builders. Extracting them into draft.fixtures.ts shrinks the
test file, makes its actual assertions easier to scan, and lets future
tests reuse the builders.
